aider: Update `CrewBuilder` constructor to accept `agent_instruments` as a parameter.
# Aider chat conversation:

USER: pass `agent_instruments` as a parameter in the constructor from

ASSISTANT: To clarify your request, could you please specify which class or method should be updated to pass `agent_instruments` as a parameter in the constructor?

USER: pass `agent_instruments` as a parameter in the constructor in `CrewBuilder`

ASSISTANT: Here is the diff to update the `CrewBuilder` class to accept `agent_instruments` as a parameter in the constructor:

```diff
--- postgres_da_ai_agent/crew_builder.py
+++ postgres_da_ai_agent/crew_builder.py
@@ ... @@
 class CrewBuilder:
-    def __init__(self):
+    def __init__(self, agent_instruments: PostgresAgentInstruments):
         self.agents = []
         self.tasks = []
         self.crew = None
         self.process = None
-        self.agent_instruments = PostgresAgentInstruments()  # Property of type PostgresAgentInstruments
+        self.agent_instruments = agent_instruments  # Property of type PostgresAgentInstruments
```

This change modifies the constructor of the `CrewBuilder` class to take an instance of `PostgresAgentInstruments` as an argument and assigns it to the `agent_instruments` property.